Camera now smoothly follows the player in the y direction. Camera also zooms out relative to the height of the player (to allow the player to see the level). Added physics material to the player to make it feel less slippy and give a small amount of bounce when landing.
