+++ Assets/Scripts/CameraControl.cs.meta
