+++ UserSettings/EditorUserSettings.asset
@@ -8,6 +8,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-0:
       value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
       flags: 0
+    RecentlyUsedScenePath-1:
+      value: 22424703114646690517182a1d1c6f3331283c2f212768252320092a
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
